--- mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/layout/NoContentMeasurePolicy.kt
@@ -1,5 +1,0 @@
-package com.jakewharton.mosaic.layout
-
-public fun interface NoContentMeasurePolicy {
-	public fun NoContentMeasureScope.measure(): MeasureResult
-}
