--- mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/layout/NoContentMeasureScope.kt
@@ -1,19 +1,0 @@
-package com.jakewharton.mosaic.layout
-
-public sealed class NoContentMeasureScope {
-	public fun layout(
-		width: Int,
-		height: Int,
-	): MeasureResult {
-		return LayoutResult(width, height)
-	}
-
-	private class LayoutResult(
-		override val width: Int,
-		override val height: Int,
-	) : MeasureResult {
-		override fun placeChildren() {}
-	}
-
-	internal companion object : NoContentMeasureScope()
-}
