+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/ui/Column.kt
@@ -3,7 +3,6 @@
 package com.jakewharton.mosaic.ui
 
 import androidx.compose.runtime.Composable
-import com.jakewharton.mosaic.layout.Layout
 import kotlin.jvm.JvmName
 
 @Composable
