+++ mosaic-runtime/src/commonMain/kotlin/com/jakewharton/mosaic/ui/Static.kt
@@ -8,7 +8,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.runtime.snapshots.SnapshotStateList
-import com.jakewharton.mosaic.Node
 import kotlin.jvm.JvmName
 
 /**
