+++ mosaic-runtime/src/commonTest/kotlin/com/jakewharton/mosaic/DebugRenderingTest.kt
@@ -1,6 +1,6 @@
 package com.jakewharton.mosaic
 
-import com.jakewharton.mosaic.layout.Layout
+import com.jakewharton.mosaic.ui.Layout
 import com.jakewharton.mosaic.ui.Row
 import com.jakewharton.mosaic.ui.Static
 import com.jakewharton.mosaic.ui.Text
